Updated bower.json "main: " to reference the file with the templates embedded.
Certain tasks like grunt-injector will reference the main file defined in bower.json to determine what source file to embed in your project.  referencing src/index.js breaks this as this file does not have the templates embedded in it.  I switched the reference to the file that works, so that the correct file is included.